+++ pfm-service/src/test/java/com/alxt/pfmservice/service/AccountServiceImplTest.java
@@ -16,6 +16,7 @@ import java.math.BigDecimal;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
@@ -32,7 +33,7 @@ class AccountServiceImplTest {
     @Test
     void findAllAccounts_FilterIsNotSet_ReturnsAccountsList() {
         // given
-        var accounts = IntStream.range(1, 4)
+        var accounts = LongStream.range(1, 4)
                 .mapToObj(i -> new Account(
                         i, "Account №%d".formatted(i), AccountType.WALLET,
                         Currency.RUB, BigDecimal.valueOf(1000), BigDecimal.valueOf(1000)))
@@ -53,7 +54,7 @@ class AccountServiceImplTest {
     @Test
     void findAllAccounts_FilterIsSet_ReturnsFilteredAccountsList() {
         // given
-        var accounts = IntStream.range(1, 4)
+        var accounts = LongStream.range(1, 4)
                 .mapToObj(i -> new Account(
                         i, "Account №%d".formatted(i), AccountType.WALLET,
                         Currency.RUB, BigDecimal.valueOf(1000), BigDecimal.valueOf(1000)))
@@ -75,21 +76,21 @@ class AccountServiceImplTest {
     void findAccount_AccountExists_ReturnsNotEmptyOptional() {
         // given
         var account = new Account(
-                1, "Account 1", AccountType.WALLET,
+                1L, "Account 1", AccountType.WALLET,
                 Currency.RUB, BigDecimal.valueOf(1000), BigDecimal.valueOf(1000)
         );
 
-        doReturn(Optional.of(account)).when(accountRepository).findById(1);
+        doReturn(Optional.of(account)).when(accountRepository).findById(1L);
 
         // when
-        var result = accountService.findAccount(1);
+        var result = accountService.findAccount(1L);
 
         // then
         assertNotNull(result);
         assertTrue(result.isPresent());
         assertEquals(account, result.orElseThrow());
 
-        verify(accountRepository).findById(1);
+        verify(accountRepository).findById(1L);
         verifyNoMoreInteractions(accountRepository);
     }
 
@@ -97,18 +98,18 @@ class AccountServiceImplTest {
     void findAccount_AccountDoesNotExist_ReturnsEmptyOptional() {
         // given
         var account = new Account(
-                1, "Account 1", AccountType.WALLET,
+                1L, "Account 1", AccountType.WALLET,
                 Currency.RUB, BigDecimal.valueOf(1000), BigDecimal.valueOf(1000)
         );
 
         // when
-        var result = accountService.findAccount(1);
+        var result = accountService.findAccount(1L);
 
         // then
         assertNotNull(result);
         assertTrue(result.isEmpty());
 
-        verify(accountRepository).findById(1);
+        verify(accountRepository).findById(1L);
         verifyNoMoreInteractions(accountRepository);
     }
 
@@ -121,7 +122,7 @@ class AccountServiceImplTest {
         );
 
         doReturn(new Account(
-                1, "Account 1", AccountType.WALLET,
+                1L, "Account 1", AccountType.WALLET,
                 Currency.RUB, BigDecimal.valueOf(1000), BigDecimal.valueOf(1000))
         ).when(accountRepository).save(payload.toAccount());
 
@@ -129,7 +130,7 @@ class AccountServiceImplTest {
         var result = accountService.createAccount(payload);
 
         // then
-        assertEquals(new Account(1, "Account 1", AccountType.WALLET,
+        assertEquals(new Account(1L, "Account 1", AccountType.WALLET,
                 Currency.RUB, BigDecimal.valueOf(1000), BigDecimal.valueOf(1000)), result);
 
         verify(accountRepository).save(new Account(null, "Account 1", AccountType.WALLET,
@@ -140,9 +141,9 @@ class AccountServiceImplTest {
     @Test
     void updateAccount_AccountExists_UpdatesAccount() {
         // given
-        var accountId = 1;
+        var accountId = 1L;
         var account = new Account(
-                1, "Account 1", AccountType.WALLET,
+                1L, "Account 1", AccountType.WALLET,
                 Currency.RUB, BigDecimal.valueOf(1000), BigDecimal.valueOf(1000)
         );
         UpdateAccountPayload payload = new UpdateAccountPayload(
@@ -151,7 +152,7 @@ class AccountServiceImplTest {
         );
 
         doReturn(Optional.of(account))
-                .when(accountRepository).findById(1);
+                .when(accountRepository).findById(1L);
 
         // when
         accountService.updateAccount(accountId, payload);
@@ -164,7 +165,7 @@ class AccountServiceImplTest {
     @Test
     void updateAccount_AccountDoesNotExist_ThrowsNoSuchElementException() {
         // given
-        var accountId = 1;
+        var accountId = 1L;
         UpdateAccountPayload payload = new UpdateAccountPayload(
                 "Account 11", AccountType.WALLET,
                 Currency.RUB, BigDecimal.valueOf(2000)
@@ -182,7 +183,7 @@ class AccountServiceImplTest {
     @Test
     void deleteAccount_DeletesAccount() {
         // given
-        var accountId = 1;
+        var accountId = 1L;
 
         // when
         accountService.deleteAccount(accountId);
